Functional not cleaned no tests

+++ src/main/java/com/example/jshop/service/AdminService.java
@@ -3,6 +3,8 @@ package com.example.jshop.service;
 import com.example.jshop.domain.category.Category;
 import com.example.jshop.domain.category.CategoryDto;
 import com.example.jshop.domain.category.CategoryWithProductsDto;
+import com.example.jshop.domain.order.Order;
+import com.example.jshop.domain.order.OrderDtoToCustomer;
 import com.example.jshop.domain.product.Product;
 import com.example.jshop.domain.product.ProductDto;
 import com.example.jshop.domain.product.ProductDtoAllInfo;
@@ -10,16 +12,16 @@ import com.example.jshop.domain.warehouse.Warehouse;
 import com.example.jshop.domain.warehouse.WarehouseDto;
 import com.example.jshop.exception.*;
 import com.example.jshop.mapper.CategoryMapper;
+import com.example.jshop.mapper.OrderMapper;
 import com.example.jshop.mapper.ProductMapper;
 import com.example.jshop.mapper.WarehouseMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
+import java.math.BigDecimal;
+import java.sql.SQLException;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class AdminService {
@@ -42,6 +44,12 @@ public class AdminService {
     @Autowired
     private WarehouseMapper warehouseMapper;
 
+    @Autowired
+    OrderService orderService;
+
+    @Autowired
+    OrderMapper orderMapper;
+
 
     public void addNewCategory(CategoryDto categoryDto) throws InvalidArgumentException, CategoryExistsException {
         Category category = categoryMapper.mapToCategory(categoryDto);
@@ -62,7 +70,10 @@ public class AdminService {
         return categoryMapper.mapToCategoryDtoAllInfo(category);
     }
 
-    public ProductDto addProduct(ProductDto productDto) throws InvalidArgumentException, CategoryExistsException {
+    public ProductDtoAllInfo addProduct(ProductDto productDto) throws InvalidArgumentException, CategoryExistsException, SQLException, InvalidPriceException {
+        if (productDto.getPrice().compareTo(BigDecimal.ZERO) < 0) {
+            throw new InvalidPriceException();
+        }
         Product product = productMapper.mapToProduct(productDto);
         Category category = categoryService.findByName(product.getCategory().getName());
         if (category != null) {
@@ -78,10 +89,13 @@ public class AdminService {
         }
         categoryService.save(category);
         Product savedProduct = productService.saveProduct(product);
-        return productMapper.mapToProductDto(savedProduct);
+        return productMapper.mapToProductDtoAllInfo(savedProduct);
     }
 
-    public ProductDto updateProduct(Long productId, ProductDto productDto) throws ProductNotFoundException, CategoryNotFoundException, InvalidArgumentException, CategoryExistsException {
+    public ProductDtoAllInfo updateProduct(Long productId, ProductDto productDto) throws ProductNotFoundException, InvalidArgumentException, CategoryExistsException, InvalidPriceException {
+        if (productDto.getPrice().compareTo(BigDecimal.ZERO) < 0){
+            throw new InvalidPriceException();
+        }
         Product productToUpdate = productService.findProductById(productId);
         Category categoryToUpdate = categoryService.findByName(productDto.getCategoryName());
         if (categoryToUpdate == null) {
@@ -95,7 +109,7 @@ public class AdminService {
         categoryService.save(categoryToUpdate);
         productToUpdate.setPrice(productDto.getPrice());
         productService.saveProduct(productToUpdate);
-        return productMapper.mapToProductDto(productToUpdate);
+        return productMapper.mapToProductDtoAllInfo(productToUpdate);
     }
 
 
@@ -124,20 +138,29 @@ public class AdminService {
         if (warehouse == null) {
             warehouse = new Warehouse(product, productQuantity);
         } else {
-            warehouse.setProductQuantity(productQuantity);
+            warehouse.setProductQuantity(warehouse.getProductQuantity() + productQuantity);
         }
         warehouseService.save(warehouse);
         return warehouseMapper.mapToWarehouseDto(warehouse);
     }
 
     public void deleteProductFromWarehouse(Long productId) throws ProductNotFoundException {
-        warehouseService.findItemByID(productId);
+        if (warehouseService.findItemByID(productId) == null){
+            throw new ProductNotFoundException();
+        };
         warehouseService.deleteById(productId);
     }
 
     public List<WarehouseDto> displayAllItemsInWarehouse() {
         List<Warehouse> listOfAllItems = warehouseService.findAll();
         return warehouseMapper.mapToWarehouseDtoList(listOfAllItems);
+    }
+
+
+    public List<OrderDtoToCustomer> displayOrders(String order_status) {
+       List<Order> listOfOrders = orderService.findOrders(order_status);
+        return orderMapper.mapToOrderDtoToCustomerList(listOfOrders);
+
     }
 }
